+++ src/events/OrderFieldEvent.php
@@ -5,6 +5,8 @@ use yii\base\Event;
 
 class OrderFieldEvent extends Event
 {
+    const FIELD_ORDER_NUMBER = 'OrderNumber';
+    const FIELD_SHIPPING_METHOD = 'ShippingMethod';
     const FIELD_CUSTOM_FIELD_1 = 'CustomField1';
     const FIELD_CUSTOM_FIELD_2 = 'CustomField2';
     const FIELD_CUSTOM_FIELD_3 = 'CustomField3';